requirements.txt
Django==3.1.dev20200423163332
djangorestframework>=3.11.0